Text Rendering
Main game creates a texture shader program using the textVertex, and textFragment shaders, and creates and calls textRenderer to render text to screen. ShaderProgram has new setUniform to send a vec3, and update textVertex and textFragment to properly render text. Currently, program just displays the glyph atlas to the screen.